fix download

+++ shell/download.sh
@@ -77,33 +77,52 @@ if [ $type == "gdrive" ]
 then
 
 	echo "${slug} Download Gdrive"
-	wget --load-cookies ${save_path}/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=${source}' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=${source}" -O ${file_save} >> ${tmp_download} 2>&1
-	rm -rf ${save_path}/cookies.txt
+	#wget --load-cookies ${save_path}/cookies.txt "https://docs.google.com/uc?export=download&confirm=$(wget --quiet --save-cookies /tmp/cookies.txt --keep-session-cookies --no-check-certificate 'https://docs.google.com/uc?export=download&id=${source}' -O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/\1\n/p')&id=${source}" -O ${file_save} >> ${tmp_download} 2>&1
+	#rm -rf ${save_path}/cookies.txt
 
-	if [ -f "$file_save" ]; then
-		echo "${slug} Download Gdrive Done"
-	else 
-		echo "${slug} Redownload Download Gdrive"
+	#if [ -f "$file_save" ]; then
+	#	echo "${slug} Download Gdrive Done"
+	#else 
+	#	echo "${slug} Redownload Download Gdrive"
+    	cd ${save_path} && sudo -u root gdrive download ${source} >> ${tmp_download} 2>&1
 		sleep 2
 		curl -sS "http://127.0.0.1:8888/rename?slug=${slug}&gid=${source}"
-    	cd ${save_path} && sudo -u root gdrive download ${source} >> ${tmp_download} 2>&1
 		sleep 1
-	fi
+	#fi
+
+	#เช็คไฟล์ download
 else
 	echo "${slug} Download Direct"
     axel -n ${speed} -o "${tmp_file}" "${source}" >> ${tmp_download} 2>&1
 fi
-sleep 1
-echo "${slug} Upload Gdrive"
-sudo -u root gdrive upload --parent ${folder} ${file_save} >> ${tmp_upload} 2>&1
-sleep 2
-echo "${slug} Save Backup"
-curl -sS "http://127.0.0.1:8888/download/backup?slug=${slug}&quality=default&sv_ip=${localip}"
-sleep 2
-echo "${slug} Done"
-curl -sS "http://127.0.0.1:8888/download/done?slug=${slug}"
-sleep 3
-rm -rf ${save_path}
-sleep 2
+
+check_api="http://127.0.0.1:8888/download/check?slug=${slug}"
+check_data=$(curl -sS "$check_api")
+check_status=$(echo $check_data | jq -r '.status')
+check_msg=$(echo $gdrive_data | jq -r '.msg')
+
+if [ $check_status == "false" ]
+then
+	echo "${slug} ${check_msg}"
+	err_api="http://127.0.0.1:8888/download/error?slug=${slug}&e_code=${error_code}&sv_ip=${localip}"
+	curl -sS "${err_api}"
+	sleep 2
+	curl -sS "http://127.0.0.1:8888/download/start?sv_ip=${localip}"
+	echo "exit"
+	exit 1
+else
+	sleep 1
+	echo "${slug} Upload Gdrive"
+	sudo -u root gdrive upload --parent ${folder} ${file_save} >> ${tmp_upload} 2>&1
+	sleep 2
+	echo "${slug} Save Backup"
+	curl -sS "http://127.0.0.1:8888/download/backup?slug=${slug}&quality=default&sv_ip=${localip}"
+	sleep 2
+	echo "${slug} Done"
+	curl -sS "http://127.0.0.1:8888/download/done?slug=${slug}"
+	sleep 3
+	rm -rf ${save_path}
+	sleep 2
+fi
 #curl -sS "http://127.0.0.1:8888/download/start?sv_ip=${localip}"
 exit 1
